Fix import issues
* Moved assertLoadingType into its own file to prevent import collision
* Moved loadActionIdentifier into its own file to prevent import collision
