+++ src/Bladeburner/Actions/utils.ts
@@ -6,30 +6,10 @@ import type { GeneralAction } from "./GeneralAction";
 
 import { BladeActionType } from "@enums";
 import { getEnumHelper } from "../../utils/EnumHelper";
-import { assertLoadingType } from "../../utils/JSONReviver";
 import { Contracts, initContracts } from "../data/Contracts";
 import { Operations, initOperations } from "../data/Operations";
 import { BlackOperations } from "../data/BlackOperations";
 import { GeneralActions } from "../data/GeneralActions";
-
-/** Returns null if the identifier is not valid */
-export function loadActionIdentifier(identifier: unknown): ActionIdentifier | null {
-  if (!identifier || typeof identifier !== "object") return null;
-  assertLoadingType<ActionIdentifier>(identifier);
-  if (getEnumHelper("BladeBlackOpName").isMember(identifier.name)) {
-    return { type: BladeActionType.blackOp, name: identifier.name };
-  }
-  if (getEnumHelper("BladeContractName").isMember(identifier.name)) {
-    return { type: BladeActionType.contract, name: identifier.name };
-  }
-  if (getEnumHelper("BladeOperationName").isMember(identifier.name)) {
-    return { type: BladeActionType.operation, name: identifier.name };
-  }
-  if (getEnumHelper("BladeGeneralActionName").isMember(identifier.name)) {
-    return { type: BladeActionType.general, name: identifier.name };
-  }
-  return null;
-}
 
 /** Return the action based on an ActionIdentifier, discriminating types when possible */
 export function getActionObject(actionId: ActionIdentifier & { type: BladeActionType.blackOp }): BlackOperation;
