+++ src/Company/Companies.ts
@@ -1,7 +1,8 @@
 // Constructs all CompanyPosition objects using the metadata in data/companypositions.ts
 import { getCompaniesMetadata } from "./data/CompaniesMetadata";
 import { Company } from "./Company";
-import { Reviver, assertLoadingType } from "../utils/JSONReviver";
+import { Reviver } from "../utils/JSONReviver";
+import { assertLoadingType } from "../utils/TypeAssertion";
 import { CompanyName } from "./Enums";
 import { PartialRecord, createEnumKeyedRecord } from "../Types/Record";
 import { getEnumHelper } from "../utils/EnumHelper";
